Tried to add alert dialogs.
The idea was to create alerts for when a badge image was clicked. It
would pull from asset/xml/badges.xml data on the requirements each badge
needed in order to be completed.

It does not work at all. I'm getting a Null Pointer Exception error when
I try to navigate from the login to the Available Badges page.

So everything I did is commented out and it works as it originally did.
It's pulling the information from the resources page that Demetri added.